chore: use a base webpack config
