+++ package.json
@@ -40,12 +40,14 @@
     "typescript": "3.3.3333",
     "validate-commit-msg": "2.14.0",
     "webpack": "4.29.5",
-    "webpack-cli": "3.2.3"
+    "webpack-cli": "3.2.3",
+    "webpack-merge": "4.2.1"
   },
   "scripts": {
     "check-types": "lerna run check-types",
     "build:types": "lerna run build:types",
-    "build": "webpack --config config/webpack.config.js",
+    "build:default": "webpack --config config/webpack.config.js",
+    "build": "lerna run build -- -d && yarn build:default -d",
     "test": "jest --coverage",
     "watch:test": "jest --watch",
     "report-coverage": "codecov",
